Move XML-related functions to a xml submodule
Ref #46
