+++ eidas_node/tests/test_saml.py
@@ -3,16 +3,15 @@ from datetime import datetime
 from pathlib import Path
 from typing import Any, BinaryIO, TextIO, cast
 
-import xmlsec
 from django.test import SimpleTestCase
 from lxml.etree import Element, ElementTree, SubElement
 
 from eidas_node.constants import LevelOfAssurance, StatusCode, SubStatusCode
 from eidas_node.errors import ValidationError
 from eidas_node.models import LightRequest, LightResponse, Status
-from eidas_node.saml import NAMESPACES, Q_NAMES, SAMLRequest, SAMLResponse, create_attribute_elm_attributes, decrypt_xml
+from eidas_node.saml import NAMESPACES, Q_NAMES, SAMLRequest, SAMLResponse, create_attribute_elm_attributes
 from eidas_node.tests.test_models import FAILED_LIGHT_RESPONSE_DICT, LIGHT_REQUEST_DICT, LIGHT_RESPONSE_DICT
-from eidas_node.utils import dump_xml, parse_xml
+from eidas_node.xml import dump_xml, parse_xml
 
 DATA_DIR = Path(__file__).parent / 'data'  # type: Path
 
@@ -35,45 +34,6 @@ class ValidationErrorMixin:
     def assert_validation_error(self, path: str, message: str, *args, **kwargs) -> Any:
         message = str(dict([(path, message)]))
         return cast(SimpleTestCase, self).assertRaisesMessage(ValidationError, message, *args, **kwargs)
-
-
-class TestDecrypt(SimpleTestCase):
-    KEY_FILE = str(DATA_DIR / 'key.pem')
-    WRONG_KEY_FILE = str(DATA_DIR / 'wrong-key.pem')
-
-    def test_decrypt_xml_with_document_not_encrypted(self):
-        with cast(BinaryIO, (DATA_DIR / 'saml_response.xml').open('rb')) as f:
-            document = parse_xml(f.read())
-        expected = dump_xml(document).decode('utf-8')
-        decrypt_xml(document, self.KEY_FILE)
-        actual = dump_xml(document).decode('utf-8')
-        self.assertXMLEqual(expected, actual)
-
-    def test_decrypt_xml_with_document_encrypted(self):
-        self.maxDiff = None
-        with cast(BinaryIO, (DATA_DIR / 'saml_response_decrypted.xml').open('rb')) as f:
-            document_decrypted = parse_xml(f.read())
-        with cast(BinaryIO, (DATA_DIR / 'saml_response_encrypted.xml').open('rb')) as f:
-            document_encrypted = parse_xml(f.read())
-        expected = dump_xml(document_decrypted).decode('utf-8')
-        decrypt_xml(document_encrypted, self.KEY_FILE)
-        actual = dump_xml(document_encrypted).decode('utf-8')
-        self.assertXMLEqual(expected, actual)
-
-    def test_decrypt_xml_with_document_encrypted_wrong_key(self):
-        self.maxDiff = None
-        with cast(BinaryIO, (DATA_DIR / 'saml_response_encrypted.xml').open('rb')) as f:
-            document_encrypted = parse_xml(f.read())
-        self.assertRaises(xmlsec.Error, decrypt_xml, document_encrypted, self.WRONG_KEY_FILE)
-
-    def test_decrypt_xml_with_document_decrypted(self):
-        self.maxDiff = None
-        with cast(BinaryIO, (DATA_DIR / 'saml_response_decrypted.xml').open('rb')) as f:
-            document_decrypted = parse_xml(f.read())
-        expected = dump_xml(document_decrypted).decode('utf-8')
-        decrypt_xml(document_decrypted, self.KEY_FILE)
-        actual = dump_xml(document_decrypted).decode('utf-8')
-        self.assertXMLEqual(expected, actual)
 
 
 class TestSAMLRequest(ValidationErrorMixin, SimpleTestCase):
